feat: add multiple new optional arguments
Also
- improve dependencies check
- complete README

+++ create/command.py
@@ -8,23 +8,26 @@ from create.app_module_files import create as create_app_module
 from create.domain_module_files import create as create_domain_module
 from create.data_module_files import create as create_data_module
 import os
+from utils.args_utils import get_project_name, get_resolved_location
+
+from utils.error_utils import raise_and_clean
 
 
 def invoke(args: dict):
-    print(args)
+    if args['--verbose']:
+        print(args)
 
-    project_name = str(args['<project>'])
+    project_name = get_project_name(args)
+    location = get_resolved_location(args)
 
-    location = args['<location>']
-    if not location:
-        location = os.getcwd()
-    location = os.path.abspath(location)
-    location += '/' + project_name.replace(' ', '-')
     print(f"🔥 Creating Android project '{project_name}' in '{location}'...")
     if os.path.isdir(location):
-        print(f"👀 The path '{location}' already exists, deleting it...")
-        shutil.rmtree(location)
-        print(f"✅️ Obsolete path '{location}' deleted")
+        override = args['--force']
+        if not override:
+            raise RuntimeError(f"The path '{location}' already exists")
+        else:
+            print(f"👀 Overriding path '{location}'")
+            shutil.rmtree(location)
 
     os.mkdir(location)
